modify: provider컴포넌트명을 provider로 끝나게 변경
