+++ src/app/_components/Map/GeocoderProvider.tsx
@@ -1,11 +1,11 @@
 'use client';
 
 import { ReactNode, createContext, useContext, useEffect, useState } from 'react';
-import { MapContext } from './Map';
+import { MapContext } from './MapProvider';
 
 export const geocoderContext = createContext<kakao.maps.services.Geocoder | undefined>(undefined);
 
-const Geocoder = ({ children }: { children: ReactNode }) => {
+const GeocoderProvider = ({ children }: { children: ReactNode }) => {
   const map = useContext(MapContext);
   const [geocoder, setGeocoder] = useState<kakao.maps.services.Geocoder>();
 
@@ -20,4 +20,4 @@ const Geocoder = ({ children }: { children: ReactNode }) => {
   return <geocoderContext.Provider value={geocoder}>{children}</geocoderContext.Provider>;
 };
 
-export default Geocoder;
+export default GeocoderProvider;
